requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
google-genai
pillow